requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
ecdsa==0.18.0
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.7.0
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
types-passlib==1.7.7.12
types-pyasn1==0.4.0.6
types-python-jose==3.3.4.8
typing_extensions==4.7.1
uvicorn==0.23.1
